+++ src/win/getaddrinfo.c
@@ -123,8 +123,8 @@ static void uv__getaddrinfo_done(struct uv__work* w, int status) {
     /* first calculate required length */
     addrinfow_ptr = req->addrinfow;
     while (addrinfow_ptr != NULL) {
-      addrinfo_len += addrinfo_struct_len +
-          ALIGNED_SIZE(addrinfow_ptr->ai_addrlen);
+      addrinfo_len += (int)(addrinfo_struct_len +
+          ALIGNED_SIZE(addrinfow_ptr->ai_addrlen));
       if (addrinfow_ptr->ai_canonname != NULL) {
         name_len = uv_utf16_to_utf8(addrinfow_ptr->ai_canonname, -1, NULL, 0);
         if (name_len == 0) {
